Allow logging directly from Game.
This is accomplished by passing in FnMut(String) listeners that log whatever string they're called with. The Log component actually has no dependency on Game so by bypassing the standard Draw and Redraw traits we can eliminate what would otherwise be a circular reference.

Unfortunately, this probably won't generalize to other sorts of listeners, which do depend on Game.