Turn MV into MVC and make a bunch of stuff way better
This is a major rearchitecting of the program. Model (Game) and View (UI) already existed. This introduces Controller in the form of UI modes. These modes run with mutable references to both the UI and the Game, as well as to the current Mode. This makes it possible to model the game flow as a graph between modes.

This includes a number of other improvements, as well as a regression or two.

Improvements:
* Introduced Examine mode to allow inspection of the map tiles. It doesn't quite work yet but it's pretty far along.
* Made movement animations look decent
* Avoided lots of full redraws of the map
* Theoretically introduced combat animations, though these don't seem to be showing up yet
* Improved the output of Game::move_unit to better capture the route of the unit and any combat along the way
* It's now possible to capture cities and start producing new units there!
* Cleaned and streamlined log output
* Drop a bunch of commented code and other cruft

Regressions:
* Fog of war is updating a little late