Units now have individualized, randomly-generated (American) names.
This introduces the `name` module which is responsible for name generation. It takes over responsibility of naming cities, and introduces functionality for naming units. The approach for unit names for now is to use U.S. Census name frequency data. In the future it would be great to broaden this to many other types of names.

A future goal for `name` is to introduce a name generation algorithm that synthesizes completely novel names based on training data.