Eliminate Redraw.
Having Redraw separate from Draw was confusing. It meant the caller always had to keep track of whether the component had been drawn already or not. The new policy is to get rid of Redraw but to make Draw track any state necessary for intelligent redraws.

This plus a few other little changes eliminate the flickering previously seen on redraws. Unit movement also doesn't recenter after every move but only on a new unit, which adds to the smoothness.

The only real downside is that moving units with large view distances is a bit slow. I think this is due to cloning every tile we draw so we can avoid redraws in the future.