Massively rework how color is handled. Initial working version, but issues remain.
We now actively think about color depth, with support for 16 color, 256 color, and true color terminals, 256 colors being the default.

We also have a method for generating player colors in an intelligent way. This comes from the `pastel` crate, or at least my fork of it.

In order to avoid a conflict with pastel's `clap` dependency, we're also upgrading `clap` to 2.33.0. This has no apparent effect on us.

TermUI is now parameterized by `C:Color+Copy` with all that entails.

This commit also includes some warning/lint fixes.