Further refine the color system.
This is largely about abstracting the notion of a color in the palette from its actual representation, whether Ansi or Rgb, or light or dark.

Most pieces of the ui now carry around a copy of the current palette.

All direct `termion` references within `game` and related are now gone.