Umpire 0.2.0: refactored, self-contained, better-looking
The refactoring is to move `unit` into `game` in preparation for moving `map`.

The self-containedness is to use `include_bytes!` to include all resources in the binary itself at compile-time. This simplifies a lot of code and averts a number of possible errors, but adds a `flate2` dependency for gunzipping the data, and updates `csv` to 1.1.1.
The release binary is now 5069104 bytes (4.9M) up from 2.something I think.

The "better-looking" makes the default color scheme use better colors earlier. Cyan is last, thank you very much.

Ah yes, also tests are now compiling. Not all passing, but compiling. It's a start.