Initial support for audio.
This uses the CPAL audio library for cross-platform support. It adds a -q (--quiet) flag to suppress audio output. A new thread is created alongside the main thread and put in charge of audio playback. It receives "orders" over an MPSC channel. This brings CPU usage up to around 15% which seems too high, but it's not terrible.