Eliminate some duplicated code and the W:Write type parameter by putting stdout in a Box.
The -a (alternate screen) option has been forcing our code into two nearly-duplicate paths because stdout was typed statically.

By making the type dynamic, we get rid of the duplication.

The cost is that our stdout is not sitting on the heap. We'll optimize it if it becomes a performance problem.