Tear out CarryOutOrdersMode and CarryOutUnitOrdersMode
This meant making Game return enough information about orders following that the UI can animate the orders after the fact.

That turned into ripping the `LogTarget` calls out of `Game`. And also to reworking begin_turn and end_turn.

Also we deprecate a bunch of overly-powerful functions on Game because we're thinking about it right now and it's probably the right thing to do eventually.

Clean up LogArea's API.

And unfortunately introduce a regression in the informativeness of unit production log messages, to be fixed soon I hope.

All in all this commit is sprawling and chaotic, but tests are passing and architecturally things feel good. Getting the `log` stuff out of `game` really cleaned up a buch of calls and obviated the need for `DefaultUI` and `DefaultLog` which I imagine we'll get rid of soon.

The big picture seems to be that the old approach of having `Game` interface with the UI via `LogTarget` and `MoveAnimator` is going away. `Game` will just be an abstract game engine as originally intended, and will return data structures describing its state and the result of any actions of interest. The UI (or future UIs) will be solely responsible for how this appears. Not sure it really matters since there's just TermUI for now, but it feels like this cleaner separation is the way of the future.