A daring and possibly foolish attempt to create an AI thread with all that entails.
This is a massive refactor that adds multithreading to `Game` and friends beyond the simple audio and input threads already in use.

This still doesn't quite compile. It may well be a dead end. I'm checkpointing it though because it's a ridiculous amount of work.

The remaining issue is that it wants `Namer` to be `Sync`. I really don't know how to do that or whether that's a good idea.

I'd guess it is also wanting `Game` to be `Sync` which may or may not be feasible.

We'll see what happens.