Begin nerfing `MapData`
AI training has smoked out a bug in which the map _should_ have a unit indexed by ID, but doesn't.

This means the constraints of `MapData` are not being upheld, particularly that all units present should be correctly indexed by location and ID, and that none should be indexed that aren't on the map.

`MapData` currently has many escape hatches (the public _mut methods) which could be leading to this problem. It also could have an error in its own logic. I'm going to audit and cut down on some of those escape hatches while I search for the bug.