Correct a logic error in unit movement.
Previously, a unit's new observations would not be considered during movement. A static ShortestPaths was generated at the start and used all along the way, regardless of whether "facts on the ground" changed. Now we recalculate the shortest paths in light of updated observations. We also manage to completely isolate the movement code from the underlying game state: it now relies entirely on the player's observations, preventing any "leaks" of information the player should not have access to.

The need to repeatedly invoke Dijkstra's algorithm at each step of the move makes this much slower. But slow is better than broken. Perhaps we can optimize in the future by bounding the shortest paths searches---right now they span the entire map, even though no unit has enough moves to move that far in a single turn.