Fix AI training and begin a reform of Game's movement methods
The AI training bugs were due to inconsistent numbering of UmpireAction's members.

Reform of the movement methods involves merging the propose_move_ and move_ methods back together so they directly modify the game state, and then using cloning to generate the proposal in exactly the same manner as it will be finally executed.