Incorporate a reinforcement learning-based AI into Umpire.
This is the culmination of a month of work (starting 6 Mar 2020 with 730f10e37). We use the `rsrl` library to train an AI using Q-Learning and then serialize it to disk. The serialized model is then incorporated into the Umpire interface as the 'a' AI player type, alongside 'h' human and 'r' random.

There are still some unfixed bugs but this is a big moment.