Index the inputs to Game::turn_is_done
The city index comes from existing city_loc_by_id, whose values constitute a deduplicated collection of city locations.

The unit index is a ne HashSet, unit_locs, which contains the location of every tile containing at least one unit (carried units included). Of course this boils down to listing the locations of toplevel units, which are a superset of the locations of all carried units.

According to `perf stat` this gives substantial gains on the AI eval task we're currently targetting:

$ cat profiling/umpire-ai_eval_perf-stat.2020-04-18T03\:35\:13-07\:00.8dcdde5
# started on Sat Apr 18 04:23:45 2020


 Performance counter stats for 'target/debug/umpire-ai -e 10 -s 100 eval random ai/10-30_e100_s100000_a__scorefix__turnpenalty.ai':

         23,046.92 msec task-clock:u              #    1.000 CPUs utilized
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             6,057      page-faults:u             #    0.263 K/sec
    74,577,575,497      cycles:u                  #    3.236 GHz
   154,093,570,518      instructions:u            #    2.07  insn per cycle
    29,820,060,437      branches:u                # 1293.885 M/sec
        30,312,008      branch-misses:u           #    0.10% of all branches

      23.049136759 seconds time elapsed

      22.967391000 seconds user
       0.009977000 seconds sys


$ cat profiling/umpire-ai_eval_perf-stat.indexed-turn-done
# started on Sat Apr 18 04:47:17 2020


 Performance counter stats for 'target/debug/umpire-ai -e 10 -s 100 eval random ai/10-30_e100_s100000_a__scorefix__turnpenalty.ai':

         16,116.62 msec task-clock:u              #    1.000 CPUs utilized
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             6,064      page-faults:u             #    0.376 K/sec
    52,116,312,784      cycles:u                  #    3.234 GHz
    83,999,434,043      instructions:u            #    1.61  insn per cycle
    14,858,166,504      branches:u                #  921.916 M/sec
        47,638,021      branch-misses:u           #    0.32% of all branches

      16.118116028 seconds time elapsed

      16.074481000 seconds user
       0.006935000 seconds sys

Also include some scripts and results for/from working with perf for a more holistic approach to benchmarking.

Alongside the prior commit, the cycles required for this AI eval task have reduced from 126 billion to 62 billion---basically a 50% savings.

The hope is that AI training and evaluation can be made efficient enough to do much more of it---training on full-sized maps, etc.

The benefits for the UI are likely substantial as well, though I don't have an automated benchmark there yet.

According to `perf record`, the top hotspots remaining is Game::update_current_player_observations