Add a data generation option to umpire-ai eval
The idea is to directly model the probability of victory---the hard definition of the reward function where the only reward is victory. This is inspired by AlphaGo which dispenses with the Bellman approximation and models the state-action value directly, using 0 for defeat and 1 for victory.

Initial experience suggests the output is extremely sparse and adds up quickly. A 10x10 run of 100 games produced over a gigabyte of data, mostly zeroes. If we can represent this sparsely we'll probably save a lot spacewise.