Drastic refactor prepping server to run the AI
Breaks out umpire-ai and umpire-tui

umpire-ai handles the AI

umpire-tui has some low-level text UI components. umpire-client is distinct in that it handles the interface with the game and builds the pieces into a running application. This was done to give umpire-ai access to some ui elements which it was using and make the training process much nicer.